refactor: app.domain 패키지 도입
