+++ api/src/main/java/vook/server/api/app/domain/user/data/RegisterCommand.java
@@ -1,4 +1,4 @@
-package vook.server.api.app.user.data;
+package vook.server.api.app.domain.user.data;
 
 import lombok.Getter;
