+++ api/src/main/java/vook/server/api/app/domain/user/exception/NotReadyToOnboardingException.java
@@ -1,9 +1,9 @@
-package vook.server.api.app.user.exception;
+package vook.server.api.app.domain.user.exception;
 
 import vook.server.api.app.common.AppException;
 
 public class NotReadyToOnboardingException extends AppException {
-    
+
     @Override
     public String contents() {
         return "NotReadyToOnboarding";
