+++ api/src/main/java/vook/server/api/app/domain/user/repo/UserInfoRepository.java
@@ -1,4 +1,4 @@
-package vook.server.api.app.user.repo;
+package vook.server.api.app.domain.user.repo;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import vook.server.api.model.user.UserInfo;
