+++ api/src/main/java/vook/server/api/app/domain/vocabulary/exception/VocabularyLimitExceededException.java
@@ -1,4 +1,4 @@
-package vook.server.api.app.vocabulary.exception;
+package vook.server.api.app.domain.vocabulary.exception;
 
 import vook.server.api.app.common.AppException;
