+++ api/src/main/java/vook/server/api/devhelper/init/LocalInit.java
@@ -5,7 +5,7 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Component;
-import vook.server.api.app.demo.repo.DemoTermRepository;
+import vook.server.api.app.domain.demo.repo.DemoTermRepository;
 import vook.server.api.devhelper.app.InitService;
 
 @Slf4j
