+++ api/src/main/java/vook/server/api/web/auth/oauth2/VookOAuth2UserService.java
@@ -8,8 +8,8 @@ import org.springframework.security.oauth2.core.OAuth2AuthenticationException;
 import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import vook.server.api.app.user.UserService;
-import vook.server.api.app.user.data.SignUpFromSocialCommand;
+import vook.server.api.app.domain.user.UserService;
+import vook.server.api.app.domain.user.data.SignUpFromSocialCommand;
 import vook.server.api.model.user.SocialUser;
 import vook.server.api.web.auth.data.VookLoginUser;
