+++ api/src/main/java/vook/server/api/web/routes/user/UserWebService.java
@@ -3,7 +3,7 @@ package vook.server.api.web.routes.user;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import vook.server.api.app.user.UserService;
+import vook.server.api.app.domain.user.UserService;
 import vook.server.api.model.user.User;
 import vook.server.api.web.auth.data.VookLoginUser;
 import vook.server.api.web.routes.user.reqres.UserInfoResponse;
