+++ api/src/main/java/vook/server/api/web/routes/user/reqres/UserRegisterRequest.java
@@ -6,7 +6,7 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 import lombok.Data;
-import vook.server.api.app.user.data.RegisterCommand;
+import vook.server.api.app.domain.user.data.RegisterCommand;
 
 @Data
 public class UserRegisterRequest {
