+++ api/src/main/java/vook/server/api/web/routes/vocabulary/VocabularyWebService.java
@@ -3,11 +3,11 @@ package vook.server.api.web.routes.vocabulary;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import vook.server.api.app.user.UserService;
-import vook.server.api.app.vocabulary.VocabularyService;
-import vook.server.api.app.vocabulary.data.VocabularyCreateCommand;
-import vook.server.api.app.vocabulary.data.VocabularyDeleteCommand;
-import vook.server.api.app.vocabulary.data.VocabularyUpdateCommand;
+import vook.server.api.app.domain.user.UserService;
+import vook.server.api.app.domain.vocabulary.VocabularyService;
+import vook.server.api.app.domain.vocabulary.data.VocabularyCreateCommand;
+import vook.server.api.app.domain.vocabulary.data.VocabularyDeleteCommand;
+import vook.server.api.app.domain.vocabulary.data.VocabularyUpdateCommand;
 import vook.server.api.model.user.User;
 import vook.server.api.model.vocabulary.Vocabulary;
 import vook.server.api.web.auth.data.VookLoginUser;
