+++ api/src/test/java/vook/server/api/testhelper/creator/TestUserCreator.java
@@ -3,10 +3,10 @@ package vook.server.api.testhelper.creator;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
-import vook.server.api.app.user.UserService;
-import vook.server.api.app.user.data.OnboardingCommand;
-import vook.server.api.app.user.data.RegisterCommand;
-import vook.server.api.app.user.data.SignUpFromSocialCommand;
+import vook.server.api.app.domain.user.UserService;
+import vook.server.api.app.domain.user.data.OnboardingCommand;
+import vook.server.api.app.domain.user.data.RegisterCommand;
+import vook.server.api.app.domain.user.data.SignUpFromSocialCommand;
 import vook.server.api.model.user.Funnel;
 import vook.server.api.model.user.Job;
 import vook.server.api.model.user.SocialUser;
