+++ api/src/test/java/vook/server/api/testhelper/creator/TestVocabularyCreator.java
@@ -3,8 +3,8 @@ package vook.server.api.testhelper.creator;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
-import vook.server.api.app.vocabulary.VocabularyService;
-import vook.server.api.app.vocabulary.data.VocabularyCreateCommand;
+import vook.server.api.app.domain.vocabulary.VocabularyService;
+import vook.server.api.app.domain.vocabulary.data.VocabularyCreateCommand;
 import vook.server.api.model.user.User;
 import vook.server.api.model.vocabulary.Vocabulary;
