+++ api/src/test/java/vook/server/api/web/routes/user/UserWebServiceTest.java
@@ -4,8 +4,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.transaction.annotation.Transactional;
-import vook.server.api.app.user.UserService;
-import vook.server.api.app.user.exception.*;
+import vook.server.api.app.domain.user.UserService;
+import vook.server.api.app.domain.user.exception.*;
 import vook.server.api.model.user.Funnel;
 import vook.server.api.model.user.Job;
 import vook.server.api.model.user.User;
