+++ api/src/test/java/vook/server/api/web/routes/vocabulary/VocabularyWebServiceTest.java
@@ -4,9 +4,9 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.transaction.annotation.Transactional;
-import vook.server.api.app.vocabulary.exception.VocabularyLimitExceededException;
-import vook.server.api.app.vocabulary.exception.VocabularyNotFoundException;
-import vook.server.api.app.vocabulary.repo.VocabularyRepository;
+import vook.server.api.app.domain.vocabulary.exception.VocabularyLimitExceededException;
+import vook.server.api.app.domain.vocabulary.exception.VocabularyNotFoundException;
+import vook.server.api.app.domain.vocabulary.repo.VocabularyRepository;
 import vook.server.api.model.user.User;
 import vook.server.api.model.vocabulary.Vocabulary;
 import vook.server.api.testhelper.IntegrationTestBase;
